+++ CHANGELOG.md
@@ -39,6 +39,7 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 * Removed `Doctrine\ORM\QueryBuilder` ([#16]), by [@localheinz]
 * Removed `Definition\AbstractDefinition` ([#114] and [#116]), by [@localheinz]
 * Removed `Definition\FakerAwareDefinition` ([#120]), by [@localheinz]
+* Removed `FixtureFactory::provideWith()` ([#122]), by [@localheinz]
 
 [fa9c564...master]: https://github.com/ergebnis/factory-bot/compare/fa9c564...master
 
@@ -63,5 +64,6 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 [#116]: https://github.com/ergebnis/factory-bot/pull/116
 [#117]: https://github.com/ergebnis/factory-bot/pull/117
 [#120]: https://github.com/ergebnis/factory-bot/pull/120
+[#122]: https://github.com/ergebnis/factory-bot/pull/122
 
 [@localheinz]: https://github.com/localheinz